+++ NEWS.md
@@ -6,6 +6,7 @@ Breaking change:
 
 New:
 
+* `vcov.hypotheses` method.
 * `wts` is now available in `plot_predictions()`, `plot_comparisons()`, and `plot_slopes()`.
 
 # marginaleffects 0.11.2